+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.19.0] - 2026-01-07
+
+### Added
+- vCard import on "Add new person" screen - drop a .vcf file to pre-fill the form
+- New API endpoint `/prm/v1/import/vcard/parse` to parse single vCard contact data
+
 ## [1.18.0] - 2026-01-07
 
 ### Changed